⌛Refactor: 마이페이지 디자인 변경
