+++ src/assets/css/user/MyInfo.css
@@ -5,9 +5,14 @@
     width: 100%;
     height: 100%;
     z-index: -1;
-    background: radial-gradient(circle, rgba(40, 78, 128, 0.9), rgba(40, 78, 128, 0.6) 70%, rgba(40, 78, 128, 0.4)), url('path/to/your/background-image.jpg') no-repeat center center fixed;
-    background-size: cover;
-    overflow: hidden; /* 배경 밖으로 넘치는 요소 감추기 */
+    /* 배경 이미지 설정 */
+    /*background-image: url("@/assets/img/myinfo_background.jpg");*/
+    background: radial-gradient(circle, #be96b2, #a8a3c7, #9293ce) no-repeat center center;
+    background-size: 200% 200%;
+    animation: gradientAnimation 5s linear infinite;
+    display: flex;
+    justify-content: center;
+    align-items: center;
 }
 
 .waterdrops {
